all: convert Binance exchanges to TypeScript
